refactor(core): move snapshot API to a specific interface
