+++ core/src/main/java/io/github/thibaultbee/streampack/core/pipelines/inputs/VideoInput.kt
@@ -19,6 +19,7 @@ import android.content.Context
 import android.graphics.Bitmap
 import android.view.Surface
 import androidx.annotation.IntRange
+import io.github.thibaultbee.streampack.core.elements.interfaces.ISnapshotable
 import io.github.thibaultbee.streampack.core.elements.processing.video.ISurfaceProcessorInternal
 import io.github.thibaultbee.streampack.core.elements.processing.video.outputs.ISurfaceOutput
 import io.github.thibaultbee.streampack.core.elements.processing.video.outputs.SurfaceOutput
@@ -45,18 +46,13 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 import kotlinx.coroutines.withContext
-import java.io.File
-import java.io.FileOutputStream
-import java.io.OutputStream
 import java.util.concurrent.atomic.AtomicBoolean
-import kotlin.coroutines.resume
-import kotlin.coroutines.suspendCoroutine
 
 /**
  * The public interface for the video input.
  * It provides access to the video source, the video processor, and the streaming state.
  */
-interface IVideoInput {
+interface IVideoInput : ISnapshotable {
 
     /**
      * Whether the video input is streaming.
@@ -85,64 +81,6 @@ interface IVideoInput {
      * The video processor for adding effects to the video frames.
      */
     val processor: ISurfaceProcessorInternal
-
-    /**
-     * Takes a snapshot of the current video frame.
-     *
-     * The snapshot is returned as a [Bitmap].
-     *
-     * @param rotationDegrees The rotation to apply to the snapshot, in degrees. 0 means no rotation.
-     * @return The snapshot as a [Bitmap].
-     */
-    suspend fun takeSnapshot(@IntRange(from = 0, to = 359) rotationDegrees: Int = 0): Bitmap
-}
-
-/**
- * Takes a JPEG snapshot of the current video frame.
- *
- * The snapshot is saved to the specified file.
- *
- * @param filePathString The path of the file to save the snapshot to.
- * @param quality The quality of the JPEG, from 0 to 100.
- * @param rotationDegrees The rotation to apply to the snapshot, in degrees.
- */
-suspend fun IVideoInput.takeJpegSnapshot(
-    filePathString: String,
-    @IntRange(from = 0, to = 100) quality: Int = 100,
-    @IntRange(from = 0, to = 359) rotationDegrees: Int = 0
-) = takeJpegSnapshot(FileOutputStream(filePathString), quality, rotationDegrees)
-
-
-/**
- * Takes a JPEG snapshot of the current video frame.
- *
- * The snapshot is saved to the specified file.
- *
- * @param file The file to save the snapshot to.
- * @param quality The quality of the JPEG, from 0 to 100.
- * @param rotationDegrees The rotation to apply to the snapshot, in degrees.
- */
-suspend fun IVideoInput.takeJpegSnapshot(
-    file: File,
-    @IntRange(from = 0, to = 100) quality: Int = 100,
-    @IntRange(from = 0, to = 359) rotationDegrees: Int = 0
-) = takeJpegSnapshot(FileOutputStream(file), quality, rotationDegrees)
-
-/**
- * Takes a snapshot of the current video frame.
- *
- * The snapshot is saved as a JPEG to the specified output stream.
- * @param outputStream The output stream to save the snapshot to.
- * @param quality The quality of the JPEG, from 0 to 100.
- * @param rotationDegrees The rotation to apply to the snapshot, in degrees.
- */
-suspend fun IVideoInput.takeJpegSnapshot(
-    outputStream: OutputStream,
-    @IntRange(from = 0, to = 100) quality: Int = 100,
-    @IntRange(from = 0, to = 359) rotationDegrees: Int = 0
-) {
-    val bitmap = takeSnapshot(rotationDegrees)
-    bitmap.compress(Bitmap.CompressFormat.JPEG, quality, outputStream)
 }
 
 /**
@@ -436,6 +374,13 @@ internal class VideoInput(
         return newSurfaceProcessor
     }
 
+    /**
+     * Takes a snapshot of the current video frame.
+     *
+     * It starts the video source if needed.
+     *
+     * The snapshot is returned as a [Bitmap].
+     */
     override suspend fun takeSnapshot(
         @IntRange(
             from = 0,
@@ -445,14 +390,13 @@ internal class VideoInput(
         if (isReleaseRequested.get()) {
             throw IllegalStateException("Input is released")
         }
-        return startStreamForBlock {
-            suspendCoroutine { continuation ->
-                val listener = processor.snapshot(rotationDegrees)
-                try {
-                    val bitmap = listener.get()
-                    continuation.resume(bitmap)
-                } catch (e: Exception) {
-                    continuation.resumeWith(Result.failure(e))
+        return withContext(dispatcherProvider.default) {
+            sourceMutex.withLock {
+                val processor = processor as? ISnapshotable?
+                    ?: throw IllegalStateException("Processor is not a snapshotable")
+
+                startStreamForBlockUnsafe {
+                    processor.takeSnapshot(rotationDegrees)
                 }
             }
         }
@@ -605,23 +549,19 @@ internal class VideoInput(
      *
      * If the stream was already running, it will not be stopped after the block.
      */
-    private suspend fun <T> startStreamForBlock(block: suspend () -> T): T {
+    private suspend fun <T> startStreamForBlockUnsafe(block: suspend () -> T): T {
         if (isReleaseRequested.get()) {
             throw IllegalStateException("Input is released")
         }
-        return withContext(dispatcherProvider.default) {
-            sourceMutex.withLock {
-                val wasStreaming = isStreamingFlow.value
-                if (!wasStreaming) {
-                    startStreamUnsafe()
-                }
-                try {
-                    block()
-                } finally {
-                    if (!wasStreaming) {
-                        stopStreamUnsafe()
-                    }
-                }
+        val wasStreaming = isStreamingFlow.value
+        if (!wasStreaming) {
+            startStreamUnsafe()
+        }
+        return try {
+            block()
+        } finally {
+            if (!wasStreaming) {
+                stopStreamUnsafe()
             }
         }
     }
